Создание и скрытие формы добавления точки
